Fixing of TIME breakouts in MySQL
Reworks :hour and :minute breakouts to function on MySQL 5.5+ and
MariaDB 10.2+

Older versions of MySQL and newer versions of MariaDB do not accept a
date of 0000-00-00, which date_format was generating. In the cases where
we're attempting to truncate the date and normalize them all to group by
hours or minutes, this will cast the TIME to DATETIME, which will make
it possible to convert it back to the proper type after truncation.

Adds an additional two test cases for grouping on a TIME column (both
:hour and :minute)

Resolves #12846

[ci mysql]